Prevent users from sharing sensitive information
This change is not about telling the user what they can and can't
do becuase if someone wants to share their IMEI then they still can.

This change is about protecting those that are not aware that they're
sharing it. Many times I've seen a user on social media wanting to share
their uptime and they share their phone number, IMEI and other sensitive
information along with it.

While many may argue that you can't fix stupid, this is a good start!

Change-Id: I6efaa8a31f6739c5d7b1b5d59bddb09d2abf30d9
Signed-off-by: RakeshBatra <raakesh.batra@rediffmail.com>